requirements.txt
aiogram==3.13.0
fastapi==0.115.0
uvicorn==0.32.0
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
